# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /professionals/{id}:
    get:
      operationId: getPublicProfessional
      tags: [professionals]
      summary: Get one public professional profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: A public professional profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicProfessional"
        "404":
          description: No such professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /professionals/me/projects:
    get:
      operationId: listMyProfessionalProjects
      tags: [professionals]
      summary: List the signed-in professional's portfolio projects
      responses:
        "200":
          description: The signed-in professional's portfolio projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProfessionalProject"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No professional profile exists yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load portfolio projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createMyProfessionalProject
      tags: [professionals]
      summary: Add a portfolio project to the signed-in professional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProfessionalProjectBody"
      responses:
        "201":
          description: Created portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionalProject"
        "400":
          description: Invalid portfolio project data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No professional profile exists yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to create portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /professionals/me/projects/{id}:
    patch:
      operationId: updateMyProfessionalProject
      tags: [professionals]
      summary: Update one of the signed-in professional's portfolio projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfessionalProjectBody"
      responses:
        "200":
          description: Updated portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProfessionalProject"
        "400":
          description: Invalid portfolio project data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such portfolio project for the signed-in professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to update portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteMyProfessionalProject
      tags: [professionals]
      summary: Delete one of the signed-in professional's portfolio projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Portfolio project deleted
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such portfolio project for the signed-in professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to delete portfolio project
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /professionals/me/quotes:
    get:
      operationId: listMyQuoteRequests
      tags: [professionals]
      summary: List quote requests sent to the signed-in professional
      responses:
        "200":
          description: Quote requests sent to the signed-in professional
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QuoteRequest"
        "401":
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No professional profile exists yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load quote requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /professionals/{id}/quote:
    post:
      operationId: createProfessionalQuoteRequest
      tags: [professionals]
      summary: Send a quote request to a professional
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateQuoteRequestBody"
      responses:
        "201":
          description: Quote request submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "400":
          description: Invalid quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No such professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to submit quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /professionals/{id}/projects:
    get:
      operationId: listPublicProfessionalProjects
      tags: [professionals]
      summary: List a professional's public portfolio projects
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Public portfolio projects for a professional
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PublicProfessionalProject"
        "404":
          description: No such professional
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Unable to load portfolio projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PublicProfessional:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        profession:
          type: string
        category:
          type: string
        services:
          type: string
        city:
          type: string
        location:
          type: ["string", "null"]
        logoUrl:
          type: ["string", "null"]
        phone:
          type: ["string", "null"]
          description: 'Public contact number, entered by the professional on their own profile. Used for the Call now action.'
        whatsapp:
          type: ["string", "null"]
          description: 'Public WhatsApp number, entered by the professional on their own profile. Used for the WhatsApp action.'
      required:
        [id, name, profession, category, services, city, location, logoUrl, phone, whatsapp]
    CreateProfessionalProjectBody:
      type: object
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        location:
          type: string
          maxLength: 200
        imageUrl:
          type: string
          maxLength: 2048
          description: 'Optional absolute http or https image URL. Non-http(s) schemes such as javascript and data URIs are rejected.'
    UpdateProfessionalProjectBody:
      type: object
      properties:
        title:
          type: string
          minLength: 2
          maxLength: 160
        description:
          type: string
          maxLength: 2000
        location:
          type: string
          maxLength: 200
        imageUrl:
          type: string
          maxLength: 2048
          description: 'Optional absolute http or https image URL. Non-http(s) schemes such as javascript and data URIs are rejected.'
    ProfessionalProject:
      allOf:
        - $ref: "#/components/schemas/CreateProfessionalProjectBody"
        - type: object
          properties:
            id:
              type: string
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
          required: [id, createdAt, updatedAt]
    PublicProfessionalProject:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
      required: [id, title, description, location, imageUrl, createdAt]
    CreateQuoteRequestBody:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
        email:
          type: string
          format: email
          maxLength: 254
        size:
          type: string
          minLength: 2
          maxLength: 120
        details:
          type: string
          minLength: 2
          maxLength: 2000
        imageUrls:
          type: array
          maxItems: 5
          items:
            type: string
            maxLength: 2048
          description: 'Optional absolute http or https image URLs, up to five. Non-http(s) schemes such as javascript and data URIs are rejected.'
      required: [name, email, size, details]
    QuoteRequest:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        size:
          type: string
        details:
          type: string
        imageUrls:
          type: array
          items:
            type: string
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, size, details, imageUrls, createdAt]