# lib/api-zod/src/generated/types/createQuoteRequestBody.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CreateQuoteRequestBody {
  /**
     * @minLength 2
     * @maxLength 120
     */
  name: string;
  /** @maxLength 254 */
  email: string;
  /**
     * @minLength 2
     * @maxLength 120
     */
  size: string;
  /**
     * @minLength 2
     * @maxLength 2000
     */
  details: string;
  /**
     * Optional absolute http or https image URLs, up to five. Non-http(s) schemes such as javascript and data URIs are rejected.
     * @maxItems 5
     * @items.maxLength 2048
     */
  imageUrls?: string[];
}